Extarct into seperate file
